FIRST COMMIT!
After a few hours I got a super basic-does nothing todo app in pure
react. I sure hope my approach of doing the basic and then refactor to
redux pays off.

The Todo app so far looks ok-ish but it could improve a lot. will leave
the fine tuning and styles for later. Now I only want to focus in the
abslolute minimum to get the app running

in tomorrow's episode we will get react-redux into the game and if
that's not so painful as just adjusting to react, we will also host it
in firebase. I probably should read a little about it