enter redux
SO... my plan was bulshit. It was actually harder to get to convert my
react app into a react+redux one. however, this part is done and now I
can focus on hosting it on firebase. After I get that part done I'll add
the extra functionality. I think I finally got what react + redux is all
about, so for now on I'll focus in the infrastructure part, and whenever
I get a chance I'll come back and add the fun logic.