Firebase hosting and hacky navigation added
SO, The firebase part was a piece of cake. nevertheless, managing the
application flow and redirections was a complete falure for around an
hour. This is not fon if you don't understand the concepts behind
react-router and history. However, we are making some progress... Next
and last step will be to connect to firebase database and validate the
login.